+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Bugfix: Fix failure to fully clone samples that have message lists as input.
+
 ## v0.3.53 (20 December 2024)
 
 - OpenAI: Support for o1 including native tool calling and `reasoning_effort` generation option.